Modified com_poll to get reid of com_xpoll and associated modules

+++ administrator/components/com_poll/views/polls/tmpl/default.php
@@ -1,18 +1,31 @@
-<?php defined('_JEXEC') or die('Restricted access'); ?>
+<?php defined('_JEXEC') or die('Restricted access');
 
-<?php JHTML::_('behavior.tooltip'); ?>
+	JHTML::_('behavior.tooltip');
 
-<?php
-	JToolBarHelper::title(  JText::_( 'Poll Manager' ) );
-	JToolBarHelper::publishList();
-	JToolBarHelper::unpublishList();
-	JToolBarHelper::deleteList();
-	JToolBarHelper::editListX();
-	JToolBarHelper::addNewX();
+	$canDo = PollHelper::getActions('component');
+
+	JToolBarHelper::title(  JText::_( 'Poll Manager' ), 'poll.png' );
+	if ($canDo->get('core.edit.state')) 
+	{
+		JToolBarHelper::publishList();
+		JToolBarHelper::unpublishList();
+	}
+	if ($canDo->get('core.delete')) 
+	{
+		JToolBarHelper::deleteList();
+	}
+	if ($canDo->get('core.edit')) 
+	{
+		JToolBarHelper::editListX();
+	}
+	if ($canDo->get('core.create')) 
+	{
+		JToolBarHelper::addNewX();
+	}
 	JToolBarHelper::help( 'screen.polls' );
 ?>
 
-<form action="index.php?option=com_poll" method="post" name="adminForm">
+<form action="index.php?option=com_poll" method="post" name="adminForm" id="adminForm">
 <table>
 	<tr>
 		<td align="left" width="100%">
@@ -30,28 +43,31 @@
 	<table class="adminlist">
 	<thead>
 		<tr>
-			<th width="5">
+			<th>
 				<?php echo JText::_( 'NUM' ); ?>
 			</th>
-			<th width="20">
+			<th>
 				<input type="checkbox" name="toggle" value="" onclick="checkAll(<?php echo count( $this->items ); ?>);" />
 			</th>
-			<th  class="title">
+			<th class="title">
 				<?php echo JHTML::_('grid.sort',   'Poll Title', 'm.title', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
 			</th>
-			<th width="5%" align="center">
+			<th>
 				<?php echo JHTML::_('grid.sort',   'Published', 'm.published', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
 			</th>
-			<th width="5%" align="center">
+			<th>
+				<?php echo JHTML::_('grid.sort',   'Open', 'm.open', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
+			</th>
+			<th>
 				<?php echo JHTML::_('grid.sort',   'Votes', 'm.voters', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
 			</th>
-			<th width="5%" align="center">
+			<th>
 				<?php echo JHTML::_('grid.sort',   'Options', 'numoptions', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
 			</th>
-			<th width="5%" align="center">
+			<th>
 				<?php echo JHTML::_('grid.sort',   'Lag', 'm.lag', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
 			</th>
-			<th width="1%" nowrap="nowrap">
+			<th>
 				<?php echo JHTML::_('grid.sort',   'ID', 'm.id', @$this->lists['order_Dir'], @$this->lists['order'] ); ?>
 			</th>
 		</tr>
@@ -73,7 +89,15 @@
 		$link 		= JRoute::_( 'index.php?option=com_poll&view=poll&task=edit&cid[]='. $row->id );
 
 		$checked 	= JHTML::_('grid.checkedout',   $row, $i );
-		$published 	= JHTML::_('grid.published', $row, $i );
+		//$published 	= JHTML::_('grid.published', $row, $i );
+		
+		$task  = $row->published ? 'unpublish' : 'publish';
+		$class = $row->published ? 'published' : 'unpublished';
+		$alt   = $row->published ? JText::_('PUBLISHED') : JText::_('UNPUBLISHED');
+
+		$task2  = ($row->open == 1) ? 'close' : 'open';
+		$class2 = ($row->open == 1) ? 'published' : 'unpublished';
+		$alt2   = ($row->open == 1) ? JText::_('OPEN') : JText::_('CLOSED');
 	?>
 		<tr class="<?php echo "row$k"; ?>">
 			<td>
@@ -83,19 +107,40 @@
 				<?php echo $checked; ?>
 			</td>
 			<td>
-			<?php if (  JTable::isCheckedOut($this->user->get ('id'), $row->checked_out )  ) {
+			<?php if (JTable::isCheckedOut($this->user->get('id'), $row->checked_out) || !$canDo->get('core.edit')) {
 				echo $row->title;
 			} else {
 				?>
-				<span class="editlinktip hasTip" title="<?php echo JText::_( 'Edit Poll' );?>::<?php echo htmlspecialchars($row->title); ?>">
-				<a href="<?php echo $link  ?>">
-					<?php echo htmlspecialchars($row->title); ?></a></span>
+				<span class="editlinktip hasTip" title="<?php echo JText::_( 'Edit Poll' );?>::<?php echo $this->escape($row->title); ?>">
+					<a href="<?php echo $link  ?>">
+						<?php echo $this->escape($row->title); ?>
+					</a>
+				</span>
 				<?php
 			}
 			?>
 			</td>
-			<td align="center">
-				<?php echo $published;?>
+			<td>
+<?php if ($canDo->get('core.edit.state')) { ?>
+				<a class="state <?php echo $class;?>" href="index.php?option=com_poll&amp;task=<?php echo $task; ?>&amp;cid[]=<? echo $row->id; ?>&amp;<?php echo JUtility::getToken(); ?>=1" title="Set this to <?php echo $task;?>">
+					<span><?php echo $alt; ?></span>
+				</a>
+<?php } else { ?>
+				<span class="state <?php echo $class;?>">
+					<span><?php echo $alt; ?></span>
+				</span>
+<?php } ?>
+			</td>
+			<td>
+<?php if ($canDo->get('core.edit.state')) { ?>
+				<a class="state <?php echo $class2;?>" href="index.php?option=com_poll&amp;task=<?php echo $task2; ?>&amp;cid[]=<? echo $row->id; ?>&amp;<?php echo JUtility::getToken(); ?>=1" title="Set this to <?php echo $task2;?>">
+					<span><?php echo $alt2; ?></span>
+				</a>
+<?php } else { ?>
+				<span class="state <?php echo $class2;?>">
+					<span><?php echo $alt2; ?></span>
+				</span>
+<?php } ?>
 			</td>
 			<td align="center">
 				<?php echo $row->voters; ?>
